requirements.txt
Flask==2.2.2
google-cloud-datastore==2.8.2

#datetime?